Refactored data package in commons
