+++ pizzaserver-commons/src/main/java/io/github/pizzaserver/commons/data/store/DataStore.java
@@ -1,6 +1,8 @@
-package io.github.pizzaserver.commons.data;
+package io.github.pizzaserver.commons.data.store;
 
+import io.github.pizzaserver.commons.data.key.DataKey;
 import io.github.pizzaserver.commons.data.react.ActionType;
+import io.github.pizzaserver.commons.data.value.ValueProxy;
 
 import java.util.Optional;
