+++ pizzaserver-commons/src/main/java/io/github/pizzaserver/commons/data/value/ValueInterface.java
@@ -1,4 +1,4 @@
-package io.github.pizzaserver.commons.data;
+package io.github.pizzaserver.commons.data.value;
 
 public interface ValueInterface<T> {
