+++ pizzaserver-server-api/src/main/java/io/github/pizzaserver/api/entity/data/attributes/AttributeType.java
@@ -1,7 +1,7 @@
 package io.github.pizzaserver.api.entity.data.attributes;
 
 import io.github.pizzaserver.api.keychain.EntityKeys;
-import io.github.pizzaserver.commons.data.DataKey;
+import io.github.pizzaserver.commons.data.key.DataKey;
 
 import java.util.*;
 import java.util.stream.Collectors;
