+++ pizzaserver-server/src/main/java/io/github/pizzaserver/server/player/ImplPlayer.java
@@ -45,8 +45,8 @@ import io.github.pizzaserver.api.scoreboard.DisplaySlot;
 import io.github.pizzaserver.api.scoreboard.Scoreboard;
 import io.github.pizzaserver.api.utils.Location;
 import io.github.pizzaserver.api.utils.TextMessage;
-import io.github.pizzaserver.commons.data.Preprocessors;
-import io.github.pizzaserver.commons.data.ValueContainer;
+import io.github.pizzaserver.commons.data.value.Preprocessors;
+import io.github.pizzaserver.commons.data.value.ValueContainer;
 import io.github.pizzaserver.commons.utils.NumberUtils;
 import io.github.pizzaserver.server.ImplServer;
 import io.github.pizzaserver.server.entity.ImplEntity;
